+++ ClimaDaemon/Core/Clima.Core/IO/IAnalogInput.cs
@@ -5,7 +5,7 @@ namespace Clima.Core.IO
         event AnalogPinValueChangedEventHandler ValueChanged;
         IAnalogValueConverter ValueConverter { get; set; }
 
-        double Value { get; }
+        float Value { get; }
         double RawValue { get; }
     }
 }
