+++ ClimaDaemon/CoreImplementations/Clima.Core.Devices/ThyristorConverter.cs
@@ -15,7 +15,7 @@ namespace Clima.Core.Devices
             throw new System.NotImplementedException();
         }
 
-        public void SetPower(double power)
+        public void SetPower(float power)
         {
             AnalogPin.SetValue(power);
         }
